backend/requirements.txt
flask
Flask-Cors
pandas
numpy
scikit-learn
pytest
python-dotenv
google-generativeai
